requirements.txt
Flask==0.10.1
Flask-Cors==2.1.0
Flask-HTTPAuth==2.5.0
Flask-SQLAlchemy==2.0
Jinja2==2.8
MarkupSafe==0.23
SQLAlchemy==1.0.8
Werkzeug==0.10.4
arrow==0.5.0
future==0.14.3
itsdangerous==0.24
memoized-property==1.0.2
mendeley==0.3.2
oauthlib==0.7.2
python-dateutil==2.4.2
requests==2.5.1
requests-oauthlib==0.4.2
six==1.9.0
